Nice things for Winterlands start
Added a bunch of fishing / hunting trades at The Polar Exchange for different meats.

Extra biomes and layouts will generate too on this setting to improve resource availability and extend what can be made without leaving.

Added more interactions to the Bear Head and nerfed some of its stats too.

Fixed a bug with fleas that would make them despawn after a reload on a host.